requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
bzip2==1.0.8=he774522_0
ca-certificates==2023.08.22=haa95532_0
click=8.1.7=py311haa95532_0
colorama==0.4.6=py311haa95532_0
flask==2.2.2=py311haa95532_0
itsdangerous==2.0.1=pyhd3eb1b0_0
jinja2==3.1.2=py311haa95532_0
libffi==3.4.4=hd77b12b_0
markupsafe==2.1.1=py311h2bbff1b_0
openssl==3.0.12=h2bbff1b_0
pip==23.3=py311haa95532_0
python==3.11.5=he1021f5_0
setuptools==68.0.0=py311haa95532_0
sqlite==3.41.2=h2bbff1b_0
tk==8.6.12=h2bbff1b_0
tzdata==2023c=h04d1e81_0
vc==4.2=h21ff451_1
vs2015_runtime==14.27.29016=h5e58377_2
werkzeug==2.2.3=py311haa95532_0
wheel==0.41.2=py311haa95532_0
xz==5.4.2=h8cc25b3_0
zlib==1.2.13=h8cc25b3_0
asgiref==3.7.2
Django==4.2.4
mysql-connector-python==8.1.0
protobuf==4.21.12
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.5.0
flask_wtf==1.2.1    
wtforms==3.1.1
